+++ doc-examples/src/main/java/react4j/examples/arez/step1/Footer.java
@@ -6,7 +6,7 @@ import react4j.annotations.ReactComponent;
 import react4j.arez.ReactArezComponent;
 import react4j.core.BaseProps;
 import react4j.core.BaseState;
-import react4j.core.ReactElement;
+import react4j.core.ReactNode;
 import react4j.dom.events.MouseEventHandler;
 import react4j.dom.proptypes.html.AnchorProps;
 import react4j.dom.proptypes.html.BtnProps;
@@ -28,7 +28,7 @@ class Footer
 
   @Nullable
   @Override
-  protected ReactElement<?, ?> renderAsElement()
+  protected ReactNode render()
   {
     final int count = AppData.model.totalCount();
     final String activeTodoWord = "item" + ( count == 1 ? "" : "s" );