+++ todomvc/src/main/java/react4j/todomvc/TodoItem.java
@@ -16,6 +16,7 @@ import react4j.arez.ReactArezComponent;
 import react4j.core.BaseProps;
 import react4j.core.BaseState;
 import react4j.core.ReactElement;
+import react4j.core.ReactNode;
 import react4j.dom.events.FocusEventHandler;
 import react4j.dom.events.FormEvent;
 import react4j.dom.events.FormEventHandler;
@@ -183,7 +184,7 @@ class TodoItem
 
   @Nullable
   @Override
-  protected ReactElement<?, ?> renderAsElement()
+  protected ReactNode render()
   {
     final Props props = props();
     final boolean completed = props.todo.isCompleted();