+++ CHANGELOG.md
@@ -42,6 +42,8 @@
 * 💥 **\[core\]** Remove `ReactRefCallback` and replace it with `RefConsumer`.
 * 💥 **\[core\]** Remove `ReactElementChildren` class and replace usages with `ReactNode` class as they are equivalent.
 * 💥 **\[core\]** Add support for the `componentDidCatch` lifecycle method introduced in React 16.
+* 💥 **\[core\]** Removed `Component.renderAsElement()` as `ReactElement` now implements `ReactNode` and thus
+  `Component.render()` is a better alternative.
 
 ### [v0.12](https://github.com/realityforge/react4j/tree/v0.12) (2017-11-09)
 [Full Changelog](https://github.com/realityforge/react4j/compare/v0.11...v0.12)